Refactor: move `.support_side()` into `Mode`
