+++ crates/walking_engine/src/mode.rs
@@ -94,4 +94,15 @@ impl Mode {
             Mode::Catching(catching) => catching.tick(context),
         }
     }
+
+    pub fn support_side(&self) -> Option<Side> {
+        match self {
+            Mode::Standing(_) => None,
+            Mode::Starting(Starting { step })
+            | Mode::Walking(Walking { step, .. })
+            | Mode::Kicking(Kicking { step, .. })
+            | Mode::Stopping(Stopping { step, .. })
+            | Mode::Catching(Catching { step, .. }) => Some(step.plan.support_side),
+        }
+    }
 }
